change loading patches to fix 500 in wiki
