--- app/views/attachments/_links.html.erb
@@ -1,29 +1,0 @@
-<div class="attachments">
-<% for attachment in attachments %>
-<p><%= link_to_attachment_with_preview attachment, :class => 'icon icon-attachment', :download => true -%>
-  <%= h(" - #{attachment.description}") unless attachment.description.blank? %>
-  <span class="size">(<%= number_to_human_size attachment.filesize %>)</span>
-  <% if options[:deletable] %>
-    <%= link_to image_tag('delete.png'), attachment_path(attachment),
-                                         :data => {:confirm => l(:text_are_you_sure)},
-                                         :method => :delete,
-                                         :class => 'delete',
-                                         :title => l(:button_delete) %>
-  <% end %>
-  <% if options[:author] %>
-    <span class="author"><%= h(attachment.author) %>, <%= format_time(attachment.created_on) %></span>
-  <% end %>
-  </p>
-<% end %>
-<% if defined?(thumbnails) && thumbnails %>
-  <% images = attachments.select(&:thumbnailable?) %>
-  <% if images.any? %>
-  <div class="thumbnails">
-    <% images.each do |attachment| %>
-      <div><%= thumbnail_with_preview_tag(attachment) %></div>
-    <% end %>
-  </div>
-  <% end %>
-<% end %>
-</div>
-
