+++ lib/redmine_lightbox.rb
@@ -1,0 +1,12 @@
+require 'redmine_lightbox/hooks/view_layouts_base_html_head_hook'
+
+require 'redmine_lightbox/patches/attachment_patch'
+
+require 'redmine_lightbox/patches/application_controller_patch'
+require 'redmine_lightbox/patches/wiki_controller_patch'
+require 'redmine_lightbox/patches/attachments_controller_patch'
+
+require 'redmine_lightbox/patches/application_helper_patch'
+require 'redmine_lightbox/patches/issues_helper_patch'
+
+require 'redmine_lightbox/services/document_converter'
